requirements.txt
webdataset
Pillow
tarfile
glob
google-cloud-vision
pyyaml